requirements.txt
# Runtime dependencies for enex2notion
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
requests>=2.31.0
w3lib>=2.1.2
tinycss2>=1.2.1
tqdm>=4.66.1
lxml>=4.9.3

# Official Notion API client
notion-client>=2.7.0
